Weeded through the unwanted bugs and slaughtered them
now only the strongest remain to be used as a breeding stock
