+++ FranticFour/Assets/01_Scripts/Controlls/PlayerFuckedController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1704d5a7ba68afb4a89fc8b98c5d29ce
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
