+++ FranticFour/Assets/01_Scripts/Controlls/RotationController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca08b104d0ce1b6419f0971d29d87b5b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
